Some filesystem changes and README additions/rename
