--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/readme.md
@@ -1,5 +1,0 @@
-# TeamCode Module
-This module contains the source code for the robot! We don't want to edit anything in the [./FtcRobotController/](FtcRobotController/) folder, as that is the base code provided by FIRST. Instead, we will be editing code in this module.
-
-## TeleOps
-- [Manual Drive](ManualDrive.java): The main manual TeleOp used for driving the robot.
